Step 2b: using collision data
OK, we've got collision data loaded into the page, ready to use in our
map.  In Mapbox GL, data is provided by _sources_, which are then used
within _layers_ to render that data to the map.

To get to the next step, you have to add a GeoJSON source using the
newly loaded collisions data, then add a layer that uses that source.