Step 3a: switching to heatmap
So we have a lot of data on a map, but we can't make sense of it!
Fortunately, Mapbox GL has several features that help us visualize large
datasets.  The first one we'll look at is _heatmap layers_: it's easy to
configure a layer to show its data not as a series of individual points,
but as a heatmap.

To get to the next step, you have to update your layer configuration to
turn it into a heatmap.