Step 3b: heatmap color ramps
We can fix this sea of red by using Mapbox GL _style expressions_, which
let us apply data-driven styling to map layers.  We'll tweak other
things in a moment, but for now we'll focus our attention on
`heatmap-color`.

To get to the next step, you have to use style expressions to create a
nice color ramp for your heatmap.