Step 3c: intensity and radius
To help de-blobify our heatmap, we can adjust `heatmap-intensity` and
`heatmap-radius` to values that better suit our data.  There isn't a
hard and fast rule here - you'll have to experiment with your dataset
(and users!) to see what values are most useful.  To make things
interesting, we'll use `interpolate` again, this time with `zoom` to
compensate for points spreading out as we zoom in.

To get to the next step, you have to use more style expressions to set
`heatmap-intensity` and `heatmap-radius`.