Step 4a: heatmap and points!
Heatmaps are great for seeing the big picture, but not so useful for
getting a precise view of individual data points...so why not use both?

To get to the next step, you have to set up another layer for points,
then use style expressions to crossfade between the heatmap and
individual points as the user zooms in.