Step 4b: clustering
Even with our new crossfade technique in place, we're still showing a
_lot_ of visual clutter with all those points.  To handle large
point-based datasets, Mapbox GL also supports _clustering_.  We'll use
this feature to help tame visual clutter at higher zoom levels.  It'll
take a bit of work, but it's worth it!

To get to the next step, you have to add a clustered data source for
collisions, then update our points layer to use that data source.