Step 4d: visual differentiation
We could still do more to help users pick out the most important data
points.  Remember those KSI collisions?  We can use `clusterProperties`
on our clustered data source to mark clusters that have _any_ KSI
collisions in them.  Once we do that, we can use this property together
with `case` expressions to dramatically increase the prominence of KSI
collisions.

To get to the next step, you have to define the `ksiAny` cluster
property, use it to style clustered circles and text labels, and use the
`ksi` property on unclustered points to style those.