Step 4d complete: visual differentiation
And here we are!  You've taken this map all the way from a blank canvas
to something that helps users see the big picture _and_ zoom in on
details, all while drawing attention to the most important points in the
dataset.

Is there more we could do?  Definitely!  Contextual popups, interactive
filters (e.g. collisions that involve cyclists, pedestrians, etc.),
additional datasets (e.g. traffic counts, road infrastructure projects,
etc.) - all these (and more!) could be added to unlock new possibilities
for understanding this dataset.

There's also lots we could do to improve accessibility.  The red text on
red circles is _definitely_ not WCAG AA compliant: it's too small, and
the contrast isn't high enough.  The map has no non-visual alternatives
(e.g. summary text), nor does it offer ways to navigate data by
keyboard.  (This is a non-trivial problem, of course!)

Finally: we're using a subset of the data.  Because of that, we can just
load it directly into the browser and play around - but what if we were
using the full 35-year dataset with 1.5 million data points, each of
which have several dozen properties?  To handle that much data, we need
to preprocess it: pre-generate vector tiles, pre-aggregate summary
statistics, pre-cluster point datasets.  By doing this, we can scale
these same techniques to work with millions (or even billions!) of data
points.